+++ src/setupTests.js
@@ -1,5 +1,4 @@
 import '@testing-library/jest-dom'
-import React from 'react'
 import { TextEncoder, TextDecoder } from 'util'
 
 // Add TextEncoder/TextDecoder for Node environment
@@ -7,29 +6,32 @@ global.TextEncoder = TextEncoder
 global.TextDecoder = TextDecoder
 
 // Mock framer-motion for tests
-jest.mock('framer-motion', () => ({
-  motion: {
-    div: ({ children, ...props }) => {
-      const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
-      return React.createElement('div', rest, children)
+jest.mock('framer-motion', () => {
+  const React = require('react')
+  return {
+    motion: {
+      div: ({ children, ...props }) => {
+        const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
+        return React.createElement('div', rest, children)
+      },
+      button: ({ children, ...props }) => {
+        const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
+        return React.createElement('button', rest, children)
+      },
+      header: ({ children, ...props }) => {
+        const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
+        return React.createElement('header', rest, children)
+      },
+      span: ({ children, ...props }) => {
+        const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
+        return React.createElement('span', rest, children)
+      }
     },
-    button: ({ children, ...props }) => {
-      const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
-      return React.createElement('button', rest, children)
-    },
-    header: ({ children, ...props }) => {
-      const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
-      return React.createElement('header', rest, children)
-    },
-    span: ({ children, ...props }) => {
-      const { initial, animate, exit, transition, whileHover, whileTap, ...rest } = props
-      return React.createElement('span', rest, children)
-    }
-  },
-  AnimatePresence: ({ children }) => children,
-  useMotionValue: () => ({ set: jest.fn() }),
-  useTransform: () => 0,
-}))
+    AnimatePresence: ({ children }) => children,
+    useMotionValue: () => ({ set: jest.fn() }),
+    useTransform: () => 0,
+  }
+})
 
 // Mock environment variables
 process.env = {
